+++ crates/prover-clients/atlantic-service/src/client.rs
@@ -48,6 +48,7 @@ impl AtlanticClient {
         let prover_type = atlantic_params.atlantic_prover_type.clone();
 
         let client = HttpClient::builder(url.as_str())
+            .expect("Failed to create HTTP client builder")
             .default_form_data("mockFactHash", &mock_fact_hash)
             .default_form_data("proverType", &prover_type)
             .build()
@@ -56,7 +57,7 @@ impl AtlanticClient {
         let proving_layer: Box<dyn ProvingLayer> = match atlantic_params.atlantic_settlement_layer.as_str() {
             "ethereum" => Box::new(EthereumLayer),
             "starknet" => Box::new(StarknetLayer),
-            _ => panic!("proving layer not correct"),
+            _ => panic!("Invalid settlement layer: {}", atlantic_params.atlantic_settlement_layer),
         };
 
         Self { client, proving_layer }
@@ -66,7 +67,7 @@ impl AtlanticClient {
         &self,
         pie_file: &Path,
         proof_layout: LayoutName,
-        atlantic_api_key: String,
+        atlantic_api_key: impl AsRef<str>,
     ) -> Result<AtlanticAddJobResponse, AtlanticError> {
         let proof_layout = match proof_layout {
             LayoutName::dynamic => "dynamic",
@@ -76,22 +77,17 @@ impl AtlanticClient {
         let response = self
             .proving_layer
             .customize_request(
-                self.client
-                    .request()
-                    .method(Method::POST)
-                    .query_param("apiKey", &atlantic_api_key)
-                    .form_file("pieFile", pie_file, "pie.zip")
-                    .form_text("layout", proof_layout),
+                self.client.request().method(Method::POST).query_param("apiKey", atlantic_api_key.as_ref()),
             )
+            .form_file("pieFile", pie_file, "pie.zip")?
+            .form_text("layout", proof_layout)
             .send()
             .await
-            .map_err(AtlanticError::AddJobFailure)
-            .expect("Failed to add job");
+            .map_err(AtlanticError::AddJobFailure)?;
 
-        if response.status().is_success() {
-            response.json().await.map_err(AtlanticError::AddJobFailure)
-        } else {
-            Err(AtlanticError::SharpService(response.status()))
+        match response.status().is_success() {
+            true => response.json().await.map_err(AtlanticError::AddJobFailure),
+            false => Err(AtlanticError::SharpService(response.status())),
         }
     }